"Introduce income statement and handle delisted companies
The commit adds functionality to import income statement data for multiple symbols. This extends coverage and helps to maintain current and comprehensive data. Error handling was also integrated into the symbol import process to handle exceptions and continue with the remaining symbols. The delistedCompaniesImporter was introduced to manage companies that have ceased to exist. Corresponding changes were made to other associated classes and files to accommodate these new importers and changes."