Improved date conversion with Unix timestamps in the Response class
There is an issue concerning the correct conversion of date and time values in the Response class. In line 129, a new check has been implemented to assess if the string value to be altered into date format is an epoch timestamp, and if so, prepends it with '@' symbol before converting it to \DateTime. This resolves a bug where numeric string values without dashes were misinterpreted as incorrect date format, throwing an exception in cases where the string represents an epoch timestamp.