Added metaData property to the Stock class
Introducing metaData property in the Stock class significantly enhances the data structure and clarity at hand. It provides details about each attribute in the Stock class like its type, phpType, length, and several other properties. This contribution presents a streamlined way to manage, work with, and understand the different components of a stock in more detail.