Correct Database class names in comments and namespace
Corrected the class names in the comments of the StockDividend.php and QuoteShort.php files in the Portfolio project. This was to avoid confusion for future developers reading the code as the original class name and the file's class name was inconsistent.

Additionally, standardized the namespaces in the QuoteShort.php file to enhance readability and maintenance. Inconsistency in namespaces could lead to difficulty in navigating and understanding the codebase. Now all classes in the pofolio\classes\FMP\Response are under the correct namespace.