Change class name CikSearch to CIKSearch
Changed the name of the class CikSearch to CIKSearch to follow the convention that acronyms in class names should be capitalized entirely. Also, unnecessary methods for acquiring CIK as String or Integer are removed because the definition of CIK objects now directly offers these functionalities. The purpose of these changes is to improve the readability and to prevent redundancy in code.