Add AvailableEuronext class and methods in FmpApiClient
Introduced a new class, AvailableEuronext, in pofolio\classes\FMP\Response. This class extends Search and has a static URL property for fetching Euronext available symbols from the API.

Also, updated FmpApiClient has a new function getAvailableEuronext(), which creates a new instance of the AvailableEuronext class for fetching all symbols for stocks traded on Euronext exchanges. This functionality is significant for expanding the list of exchanges covered, adding one of the largest exchanges in Europe, thus providing users with a broader range of stocks to explore and trade.