Add ExchangeSymbols class and update related method
Introduced a new class, ExchangeSymbols, in 'pofolio\classes\FMP\Response' path and also updated the FmpApiClient class to include methods to handle these exchange symbols. The change aims to fetch and handle stock exchange symbols more efficiently.

In addition to this a method 'getExchangeSymbols' was added to initiate and return an instance of the ExchangeSymbols class, extending the capability of the FmpApiClient. Also, modified existing method names 'getAvailableEuronext' and 'getSymbolChange' to 'getAvailableEuronextSymbols' and 'getSymbolChanges' respectively for better clarity.