Add ExecutiveCompensationBenchmark to FMP API client
Created a new class 'ExecutiveCompensationBenchmark' and updated 'FmpApiClient' to include references to this class. This enhancement allows fetching of executive compensation benchmark data which can be used for executive pay comparison across industries for a particular year. This additional data can provide insightful context for executive pay analysis in the financial market project.